% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchconfig.R
\name{blockwatch.api_key}
\alias{blockwatch.api_key}
\title{Query or set Blockwatch API key}
\usage{
blockwatch.api_key(api_key)
}
\arguments{
\item{api_key}{Optionally passed parameter to set Blockwatch \code{api_key}.}
}
\value{
Returns invisibly the currently set \code{api_key}.
}
\description{
Query or set Blockwatch API key
}
\examples{
\dontrun{
blockwatch.api_key('foobar')
}
}
