% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchconfig.R
\name{blockwatch.base_url}
\alias{blockwatch.base_url}
\title{Query or set Blockwatch API base url}
\usage{
blockwatch.base_url(base_url)
}
\arguments{
\item{base_url}{Optionally passed parameter to set Blockwatch \code{base_url}.}
}
\value{
Returns invisibly the currently set \code{base_url}.
}
\description{
Query or set Blockwatch API base url
}
\examples{
\dontrun{
blockwatch.base_url('https://data.blockwatch.cc/v1')
}
}
