% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchsearch.R
\name{blockwatch.databases}
\alias{blockwatch.databases}
\title{Lists subscribed databases}
\usage{
blockwatch.databases(silent = FALSE)
}
\arguments{
\item{silent}{Prints the results when FALSE (default).}
}
\description{
Lists subscribed databases
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}

For instructions on finding your authentication token go to https://blockwatch.cc/account/profile#apikey
}
\examples{
\dontrun{
blockwatch.databases()
}
}
\seealso{
\code{\link{blockwatch.api_key}}
}
