% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchsearch.R
\name{blockwatch.datafields}
\alias{blockwatch.datafields}
\title{Lists datafields}
\usage{
blockwatch.datafields(code, silent = FALSE)
}
\arguments{
\item{silent}{Prints the results when FALSE (default).}

\item{dataset_code}{Fully qualified Blockwatch dataset code (DB_CODE/SET_CODE).}
}
\description{
Lists datafields
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}

For instructions on finding your authentication token go to https://blockwatch.cc/account/profile#apikey
}
\examples{
\dontrun{
blockwatch.datafields("BITFINEX:OHLCV/BTC_USD")
}
}
\seealso{
\code{\link{blockwatch.api_key}}
}
