% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchsearch.R
\name{blockwatch.datasets}
\alias{blockwatch.datasets}
\title{Lists datasets}
\usage{
blockwatch.datasets(database_code, silent = FALSE)
}
\arguments{
\item{database_code}{Blockwatch database code.}

\item{silent}{Prints the results when FALSE (default).}
}
\description{
Lists datasets
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}

For instructions on finding your authentication token go to https://blockwatch.cc/account/profile#apikey
}
\examples{
\dontrun{
blockwatch.datasets("BITFINEX:OHLCV")
}
}
\seealso{
\code{\link{blockwatch.api_key}}
}
