% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchdownload.R
\name{blockwatch.download}
\alias{blockwatch.download}
\title{Downloads a zip with all data from a Blockwatch database}
\usage{
blockwatch.download(database_code, filename, ...)
}
\arguments{
\item{database_code}{Database code on Blockwatch specified as a string.}

\item{filename}{Filename (including path) of file to download.}

\item{...}{Additional named values that are interpreted as Blockwatch API parameters. Please see \url{https://blockwatch.cc/docs/api} for a full list of parameters.}
}
\value{
The filename of the downloaded file.
}
\description{
Downloads a zip with all data from a Blockwatch database
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}
}
\examples{
\dontrun{
blockwatch.download("BITFINEX:OHLCV/BTC_USD", "./BITFINEX_OHLVC_BTC_USD.zip")
}
}
\seealso{
\code{\link{blockwatch.api_key}}
}
