% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchsearch.R
\name{blockwatch.help}
\alias{blockwatch.help}
\title{Help displays information about a Blockwatch database, dataset and datafields}
\usage{
blockwatch.help(code)
}
\arguments{
\item{code}{Database, dataset (DB/SET) or datafield (DB/SET/FIELD) code}
}
\description{
Help displays information about a Blockwatch database, dataset and datafields
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}

For instructions on finding your authentication token go to https://blockwatch.cc/account/profile#apikey
}
\examples{
\dontrun{
blockwatch.help("BITFINEX:OHLCV")
}
}
\seealso{
\code{\link{blockwatch.api_key}}
}
