% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchconfig.R
\name{blockwatch.quiet}
\alias{blockwatch.quiet}
\title{Disable or enable warning messages}
\usage{
blockwatch.quiet(enable = TRUE)
}
\arguments{
\item{enable}{True will disable warnings.}
}
\description{
Disable or enable warning messages
}
\examples{
\dontrun{
blockwatch.quiet()
}
}
