% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchsearch.R
\name{blockwatch.search}
\alias{blockwatch.search}
\title{Search all Blockwatch databases}
\usage{
blockwatch.search(query, silent = FALSE, per_page = 10, ...)

blockwatch.search(query, silent = FALSE, per_page = 10, ...)
}
\arguments{
\item{query}{Search terms}

\item{silent}{Prints the results when FALSE (default).}

\item{per_page}{Number of results returned per page.}

\item{...}{Additional named values that are sent as Blockwatch API parameters.}

\item{query}{Search terms}

\item{silent}{Prints the results when FALSE (default).}

\item{per_page}{Number of results returned per page.}

\item{...}{Additional named values that are sent as Blockwatch API parameters.}
}
\value{
Search results returned as a data.frame.

Search results returned as a data.frame.
}
\description{
Search all Blockwatch databases

Search all Blockwatch databases
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}

For instructions on finding your authentication token go to https://blockwatch.cc/account/profile#apikey

Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/account/profile#apikey}

For instructions on finding your authentication token go to https://blockwatch.cc/account/profile#apikey
}
\examples{
\dontrun{
search.results <- blockwatch.search("oil")
}
\dontrun{
search.results <- blockwatch.search("oil")
}
}
\seealso{
\code{\link{blockwatch.api_key}}

\code{\link{blockwatch.api_key}}
}
