% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchseries.R
\name{blockwatch.series.get}
\alias{blockwatch.series.get}
\title{Internal method to retrieve data from a Blockwatch time series endpoint}
\usage{
blockwatch.series.get(code, params)
}
\arguments{
\item{code}{Dataset code on Blockwatch specified as a string.}

\item{params}{A list of parameters to be passed to the Blockwatch API. Please see \url{https:/blockwatch.cc/docs/api#retrieve-data-and-metadata} for a full list of parameters.}
}
\value{
Returns a data.frame of the requested data
}
\description{
Internal method to retrieve data from a Blockwatch time series endpoint
}
\examples{
\dontrun{
data = blockwatch.series.get("TRADE/BITFINEX_BTC_USD", list(limit=5))
plot(data[,1])
}
}
\seealso{
\code{\link{blockwatch.api_key}}, \code{\link{blockwatch.table}}
}
