% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchseries.R
\name{blockwatch.series}
\alias{blockwatch.series}
\title{Retrieves Data from a Blockwatch Time Series endpoint}
\usage{
blockwatch.series(code, datatype = c("raw", "ts", "zoo", "xts",
  "timeSeries"), transform = c("", "diff", "rdiff", "normalize", "cumul",
  "rdiff_from"), collapse = c("", "1m", "5m", "15m", "30m", "1h", "3h",
  "6h", "12h", "1d", "1w", "1M", "3M", "1y", "daily", "weekly", "monthly",
  "quarterly", "annual"), order = c("asc", "desc"),
  force_irregular = FALSE, ...)
}
\arguments{
\item{code}{Dataset code on Blockwatch specified as a string or an array of strings.}

\item{datatype}{Type of data returned specified as string. Can be 'raw', 'ts', 'zoo', 'xts' or 'timeSeries'.}

\item{transform}{Apply Blockwatch API data transformations.}

\item{collapse}{Collapse frequency of Data.}

\item{order}{Select if data is given to R in ascending or descending formats. Helpful for the rows parameter.}

\item{force_irregular}{When set to TRUE, forces the index of the Data to be of date format yyyy-mm-dd}

\item{...}{Additional named values that are interpreted as Blockwatch API parameters. Please see \url{https:/blockwatch.cc/docs/api#database-metadata-api} for a full list of parameters.}
}
\value{
Depending on the type the class is either data.frame, time series, xts, zoo or timeSeries.
}
\description{
Retrieves Data from a Blockwatch Time Series endpoint
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https:/blockwatch.cc/account/profile#apikey}.
}
\examples{
\dontrun{
data = blockwatch.series("BITFINEX:OHLCV/BTC_USD", start_date="2018-01-01", datatype="ts")
plot(data[,1])
}
}
\references{
This R package uses the Blockwatch API. For more information go to \url{https:/blockwatch.cc/docs/api}. For more help on the package itself go to \url{https:/blockwatch.cc/docs/r}.
}
\seealso{
\code{\link{blockwatch.api_key}}
}
