% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockwatchtable.R
\name{blockwatch.table}
\alias{blockwatch.table}
\title{Retrieves Data from a Blockwatch Datatable endpoint}
\usage{
blockwatch.table(table_code, paginate = FALSE, ...)
}
\arguments{
\item{table_code}{Datatable code on Blockwatch specified as a string.}

\item{paginate}{When set to TRUE, fetches up to 1,000,000 rows of data}

\item{...}{Additional named values that are interpreted as Blockwatch API parameters.}
}
\value{
Returns a data.frame.
}
\description{
Retrieves Data from a Blockwatch Datatable endpoint
}
\details{
Set your \code{api_key} with \code{blockwatch.api_key} function. For instructions on finding your api key go to \url{https://blockwatch.cc/accountprofile#/apikey}
}
\examples{
\dontrun{
blockwatch.table('BITFINEX:TRADE/BTC_USD', paginate=TRUE)
}
}
\seealso{
\code{\link{blockwatch.api_key}}
}
